+++ tightdb-java-test/src/test/java/com/tightdb/JNIDistinctTest.java
@@ -11,7 +11,7 @@ public class JNIDistinctTest {
 
     void init() {
         table = new Table();
-        table.addColumn(ColumnType.LONG, "number");
+        table.addColumn(ColumnType.INTEGER, "number");
         table.addColumn(ColumnType.STRING, "name");
 
         long i = 0;
