+++ tightdb-java-test/src/test/java/com/tightdb/experiment/ColumnTypeViewTest.java
@@ -23,7 +23,7 @@ public class ColumnTypeViewTest {
         t  = new Table();
         t.addColumn(ColumnType.DATE, "Date");
         t.addColumn(ColumnType.STRING, "String");
-        t.addColumn(ColumnType.LONG , "Long");
+        t.addColumn(ColumnType.INTEGER , "Long");
         
         t.add(new Date(), "I'm a String", 33);
